requirements.txt
Flask==3.1.0
SQLAlchemy==2.0.0
Flask-JWT-Extended==4.7.1
bcrypt==4.0.1
PyMySQL==1.0.2  # Pour MySQL
python-dotenv==0.19.2  # Pour gérer les variables d'environnement
cryptography==44.0.0